Add TransitionHook mixin tests
