+++ tests.webpack.js
@@ -5,6 +5,7 @@ require('./modules/__tests__/Router-test');
 require('./modules/__tests__/RouteUtils-test');
 require('./modules/__tests__/RoutingUtils-test');
 require('./modules/__tests__/transitionHooks-test');
+require('./modules/__tests__/TransitionHook-test');
 require('./modules/__tests__/HashHistory-test');
 require('./modules/__tests__/BrowserHistory-test');
 require('./modules/__tests__/MemoryHistory-test');
